Move main._handle_exception test to main_test.py
